backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
neo4j==5.25.0
langchain==0.3.7
langchain-openai==0.2.9
langchain-google-genai==2.0.6
langgraph==0.2.53
beautifulsoup4==4.12.3
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2
reportlab==4.2.5
openai==1.55.3
lxml==5.3.0
langchain-ollama
